+++ index.js
@@ -1,6 +1,7 @@
 const httpClient = require('./src/httpClient');
 const AccessRecordsApi = require('./src/accessRecordsApi');
 const UserPermissionsApi = require('./src/userPermissionsApi');
+const UsersApi = require('./src/usersApi');
 const ServiceClientsApi = require('./src/serviceClientsApi');
 const ResourcesApi = require('./src/resourcesApi');
 const AccountsApi = require('./src/accountsApi');
@@ -16,6 +17,7 @@ class AuthressClient {
     this.accessRecords = new AccessRecordsApi(this.httpClient);
     this.serviceClients = new ServiceClientsApi(this.httpClient);
     this.userPermissions = new UserPermissionsApi(this.httpClient);
+    this.users = new UsersApi(this.httpClient);
     this.resources = new ResourcesApi(this.httpClient);
     this.accounts = new AccountsApi(this.httpClient);
     this.roles = new RolesApi(this.httpClient);